feature: Remove vss-processor and move Code Generation to vss-processor-plugin
