+++ samples/build.gradle.kts
@@ -3,7 +3,6 @@ import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 plugins {
     id("application")
     kotlin("jvm")
-    id("com.google.devtools.ksp") version "2.0.0-1.0.24"
     id("org.eclipse.velocitas.vss-processor-plugin") // version <VERSION>
 }
 
@@ -19,8 +18,6 @@ kotlin {
 }
 
 dependencies {
-    ksp(project(":vss-processor"))
-
     implementation(project(":sdk"))
 
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.8.1")
