+++ vss-processor-plugin/src/main/java/org/eclipse/velocitas/vssprocessor/parser/FileParseException.kt
