+++ vss-processor-plugin/src/main/java/org/eclipse/velocitas/vssprocessor/parser/json/extension/JsonObjectExtension.kt
