+++ vss-processor-plugin/src/main/java/org/eclipse/velocitas/vssprocessor/spec/VssNodePropertiesBuilder.kt
