+++ vss-processor-plugin/src/test/kotlin/org/eclipse/velocitas/vssprocessor/parser/json/JsonVssParserTest.kt
