+++ vss-processor-plugin/src/test/kotlin/org/eclipse/velocitas/vssprocessor/spec/VssNodeSpecModelTest.kt
