+++ vss-processor-plugin/src/test/resources/json/incompatible.json
