+++ vss-processor-plugin/src/test/resources/json/vss_rel_2.0.json
