+++ vss-processor-plugin/src/test/resources/json/vss_rel_3.0.json
