+++ vss-processor-plugin/src/test/resources/json/vss_rel_4.0.partial.json
