+++ vss-processor-plugin/src/test/resources/json/vss_rel_4.1.json
