+++ vss-processor-plugin/src/test/resources/yaml/incompatible.yaml
