+++ vss-processor-plugin/src/test/resources/yaml/invalid.yaml
