+++ vss-processor-plugin/src/test/resources/yaml/vss_rel_4.0.partial.yaml
