+++ vss-processor-plugin/src/test/resources/yaml/vss_rel_4.1.yaml
