--- vss-processor/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,1 +1,0 @@
-org.eclipse.velocitas.vssprocessor.VssModelGeneratorProcessorProvider
